owm-weather-provider/code/requirements.txt
kafka-python
pyowm
